+++ lib/chromic_pdf/pdf/connection.ex
@@ -43,7 +43,7 @@ defmodule ChromicPDF.Connection do
 
   defp spawn_chrome(opts) do
     opts
-    |> Keyword.take([:discard_stderr, :no_sandbox])
+    |> Keyword.take([:chrome_args, :discard_stderr, :no_sandbox])
     |> @chrome.spawn()
   end