+++ test/unit/chromic_pdf/pdf/connection_test.exs
@@ -24,7 +24,7 @@ defmodule ChromicPDF.ConnectionTest do
 
   describe "initialization" do
     test "it spawns Chrome and initializes its state" do
-      opts = [discard_stderr: false, no_sandbox: true]
+      opts = [discard_stderr: false, no_sandbox: true, chrome_args: "--foo"]
       expect(ChromeMock, :spawn, fn ^opts -> {:ok, @port} end)
       assert init({self(), opts}) == {:ok, new_state()}
     end